Allow with-match to match symbols anywhere.
Before, something like

(with-arg (sym1 "--between--" sym2

was impossible, because the first element of the list had to be
a string.

As a positive side effect of these changes, with-match now silently
ignores any bad matches (returns nil instead of throwing run-time
errors). It also no longer requires a helper function (drop-until-
string=).